+++ iOS/KHUPASS/Screen/ScannerScreen/ScannerController.swift
